requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
phonenumbers==9.0.8
sqlparse==0.5.3
tzdata==2025.2
